BB-20380: Improve performance of Merge by Priority pricing strategy (#29200)
- added hint
- fixed test and repository
- removed deprecated code
